# Conflicts:
#	src/main/java/com/example/unirideapi/controller/RutaController.java
#	src/main/java/com/example/unirideapi/repository/PasajeroRepository.java
#	src/main/java/com/example/unirideapi/service/impl/RutaServiceImpl.java

+++ src/main/java/com/example/unirideapi/service/impl/SolicitudViajeServiceImpl.java
@@ -36,7 +36,7 @@ public class SolicitudViajeServiceImpl implements SolicitudViajeService {
     @Override
     public SolicitudViajeResponseDTO create(SolicitudViajeRequestDTO solicitudViajeRequestDTO){
 
-        Pasajero pasajero = pasajeroRepository.findById((long)solicitudViajeRequestDTO.pasajeroId())
+        Pasajero pasajero = pasajeroRepository.findById(solicitudViajeRequestDTO.pasajeroId())
                 .orElseThrow(() -> new ResourceNotFoundException("Pasajero no encontrado"));
         Ruta ruta = rutaRepository.findById((long)solicitudViajeRequestDTO.rutaId())
                 .orElseThrow(() -> new ResourceNotFoundException("Pasajero no encontrado"));
